requirements.txt
fastapi
uvicorn[standard]
jinja2
pymongo
python-dotenv
